need coordinates in signal

+++ modeling.py
@@ -1,11 +1,8 @@
 from radar import Radar, Emitter, Receiver
-from objects import Aircraft
+from objects.aircraft import Aircraft
 from signal import Signal
 from collision_detector import CollisionDetector
 from coordinates import Vector3D
-
-
-
 
 
 class Scene:
@@ -19,17 +16,21 @@ class Scene:
         self.collision_detector = CollisionDetector()
 
     def update(self):
+        if self.duration <= self.time:
+            return
         self.time += 1
-        v3d = Vector3D(1, 2, 3) 
-        sp3d = Vector3D(3, 3, 3)
-        signal = self.radar.emitter.send_signal(0, self.time, v3d, sp3d)
-        self.signals.append(signal)
-        # TODO: rewrite 
-        signals_detection_object = self.collisionDetector.scan_objects(self.signals, self.objects)
+        if len(self.signals) == 0:
+            v3d = Vector3D(1, 0, 0) 
+            sp3d = Vector3D(3, 0, 0)
+            signal = self.radar.emitter.send_signal(0, self.time, v3d, sp3d)
+            self.signals.append(signal)
+        # TODO: rewrite
+        signals_detection_object = self.collision_detector.scan_objects(self.signals, self.objects)
         #update_signals(new_signals)#UpdateSignals
 
-        signals_detection_radar = self.collision_detector.scan_radar(self.signals, self.radar)
-        self.radar.receiver.process(signals_detection_radar)
-        for signal_f in signals_detection_radar:
-            self.signals.remove(signal_f)
+        # signals_detection_radar = self.collision_detector.scan_radar(self.signals, self.radar)
+        # self.radar.receiver.process(signals_detection_radar)
+        # for signal_f in signals_detection_radar:
+        #     self.signals.remove(signal_f)
+        # and remove signal which so far from radar
